refactor: change some classes to receiveprotocoldispatcher
