+++ src/Imlight.Net/MessageService.cs
@@ -11,41 +11,24 @@ using WizUnraveler.DML;
 
 namespace Imlight.Net
 {
-    public abstract class MessageService : ReceiveActor
+    public abstract class MessageService : ReceiveProtocolDispatcher
     {
         public static readonly string ASK_IDENTIFY = "IDENTIFY_YOURSELF";
 
         protected SessionActor SessionActor { get; set; }
-        public virtual Dictionary<Type, MethodInfo> MessageHandlers { get; private set; }
 
         public MessageService(SessionActor sessionActor)
         {
             this.SessionActor = sessionActor;
-
-            SetMessageHandlers();
-            ConfigureReceivers();
         }
          
-        protected virtual void ConfigureReceivers()
+        protected override void ConfigureReceivers()
         {
+            base.ConfigureReceivers();
+            
             Receive<string>(x => x == ASK_IDENTIFY, x => 
             {
                 Sender.Tell(new INTMSG_SERVICE_IDENTITY(this), Context.Self);
-            });
-
-            Receive<INetworkMessage>(message =>
-            {
-                // Find the method that handles this message type
-                if (MessageHandlers.TryGetValue(message.GetType(), out var method))
-                {
-                    // Invoke the method with the message
-                    method.Invoke(this, new object[] { message });
-                }
-                else
-                {
-                    // No handler for this message type
-                    Unhandled(message);
-                }
             });
 
             Receive<IInternalMessage>(internalMessage =>
@@ -86,11 +69,6 @@ namespace Imlight.Net
             SessionActor.GetActorRef().Tell(msg);
         }
 
-        protected void SendCloseSession()
-        {
-            SessionActor.GetActorRef().Tell("Close");
-        }
-
         protected T AskInternal<T>(IInternalMessage msg)
             where T : IInternalMessage
         {
@@ -104,30 +82,24 @@ namespace Imlight.Net
 
             return task;
         }
-
-        private void SetMessageHandlers()
+        
+        protected T AskServer<T>(INetworkMessage msg)
+            where T : INetworkMessage
         {
-            MessageHandlers = new Dictionary<Type, MethodInfo>();
-
-            // Get all methods in this actor with a message handling attribute
-            var methods = this
-                .GetType()
-                .GetMethods(BindingFlags.Instance | BindingFlags.Public | BindingFlags.NonPublic)
-                .Where(method => method.GetCustomAttribute<MessageHandlerAttribute>() != null);
-
-            if (methods.Count() <= 0)
+            if (SessionActor is null)
             {
-                Log.Logger.Warning($"{this.GetType()} does not have any methods with attribute {nameof(MessageHandlerAttribute)}." +
-                    $"Is this intended behavior?");
-
-                return;
+                Log.Logger.Error($"{this.GetType()} attempted to send message to undefined SessionActor.");
+                return default(T);
             }
 
-            foreach (var method in methods)
-            {
-                var paramType = method.GetParameters()[0].ParameterType;
-                MessageHandlers.Add(paramType, method);
-            }
+            var task = SessionActor.AskServer<T>(msg);
+
+            return task;
+        }
+        
+        protected void SendCloseSession()
+        {
+            SessionActor.GetActorRef().Tell("Close");
         }
     }
 }
