+++ src/Imlight.Net/ReceiveProtocolDispatcher.cs
@@ -1,0 +1,70 @@
+using System.Collections.Generic;
+using System.Linq;
+using System.Reflection;
+using Akka.Actor;
+using Imlight.Common;
+using WizUnraveler.DML;
+
+namespace Imlight.Net
+{
+    /// <summary>
+    /// An extension of a ReceiveActor that allows for receiving INetworkRecords directly to method attributes.
+    /// </summary>
+    public class ReceiveProtocolDispatcher : ReceiveActor
+    {
+        public virtual Dictionary<System.Type, MethodInfo> MessageHandlers { get; private set; }
+        
+        public ReceiveProtocolDispatcher()
+        {
+            SetMessageHandlers();
+            ConfigureReceivers();
+        }
+
+        protected virtual void ConfigureReceivers()
+        {
+            Receive<INetworkMessage>(message =>
+            {
+                // Find the method that handles this message type
+                if (MessageHandlers.TryGetValue(message.GetType(), out var method))
+                {
+                    // Invoke the method with the message
+                    method.Invoke(this, new object[] { message });
+                }
+                else
+                {
+                    // No handler for this message type
+                    Unhandled(message);
+                }
+            });
+        }
+        
+        private void SetMessageHandlers()
+        {
+            MessageHandlers = new Dictionary<System.Type, MethodInfo>();
+
+            // Get all methods in this actor with a message handling attribute
+            var methods = this
+                .GetType()
+                .GetMethods(
+                    BindingFlags.Instance 
+                            | BindingFlags.Public 
+                            | BindingFlags.NonPublic
+                            | BindingFlags.FlattenHierarchy)
+                .Where(method => method.GetCustomAttribute<MessageHandlerAttribute>() != null);
+
+            if (methods.Count() <= 0)
+            {
+                Log.Logger.Warning($"{this.GetType()} does not have any methods with attribute {nameof(MessageHandlerAttribute)}." +
+                                   $" Is this intended behavior?");
+
+                return;
+            }
+
+            foreach (var method in methods)
+            {
+                var paramType = method.GetParameters()[0].ParameterType;
+                MessageHandlers.Add(paramType, method);
+            }
+        }
+    }
+}
